ProgrammingForAI-taba/education_analysis/tests/requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
